Controla eliminación de usuario inv con operaciones

+++ ajax/Usuario/delUsuario.php
@@ -13,17 +13,34 @@
 		    }
 		    else
 		    {
-			   @$ID = $request->ID;  			
-				$query="delete from usuario where id=$ID";
-				if($mysqli->query($query))
-				{
-					echo $json_response = json_encode($mysqli->affected_rows);
-				}
-				else
-				{
-					 header("HTTP/1.0 404 Not Found");
-		    		 echo '{"data": '+$mysqli->error+'}';
-				}
+
+			   @$ID = $request->ID;
+
+
+                $COUNT = 0;
+                $qordenes = "SELECT COUNT(1) as contador FROM orden WHERE IdUsuario = $ID";
+                $resultado = $mysqli->query($qordenes);
+                while($row = $resultado->fetch_assoc()) {
+                        $COUNT = $row["contador"];
+                }
+
+                if ($COUNT <= 0) {
+                    $query="delete from usuario where id=$ID";
+                    if($mysqli->query($query))
+                    {
+                        echo $json_response = json_encode($mysqli->affected_rows);
+                    }
+                    else
+                    {
+                         header("HTTP/1.0 404 Not Found");
+                         echo '{"data": '+$mysqli->error+'}';
+                    }
+                } else {
+                    header("HTTP/1.0 404 Not Found");
+                    echo '{"data": "Usuario tiene ordenes existentes"}';
+                }
+
+
 				
 			}
 		}
